fixes workflow context data types to be consistent
Previously most workflow context data types were either an empty array if empty
or an object if some data was present.
This affected relations, projects and attributes,
although the bug occurred mostly with relations (they are most likely to be empty).

All of these have been fixed by replacing the array data type with a stdClass,
which will always serialize to a json object even if empty.