Fix singleEdit update of selectQuery attribute without selections
When removing all selections from a selectQuery attribute in singleEdit mode,
validate if the attribute is mandatory.

If it is mandatory then redirect to the ci detail page with a error notification
instead of displaying a broken edit box.
If it is optional then delete the attribute
instead of showing a error notification and a broken edit box.