Move password change to separate form
This fixes the issue of password managers automatically filling password fields.
If only some fields were filled it made it annoying to edit user details,
because the password fields had to be cleared manually.
If all fields were filled on the edit page of another user
this could result in a unwanted password change of said user.
Using autocomplete="off" or autocomplete="new-password" is not reliable,
to prevent autofilling of passwords.

Now a "Change password" button is offered on the user edit page,
which redirects to a new separate password change form.