Add list query to "Dropdown (SQL filled") attribute
The edit query (previously the only query for this attribute type) was only executed once in the ci list for all attributes.
The first ci id was used as placeholder and the result list was taken for the display of all other attribute values.
When the ci id was used as restriction in the query, other selected attribute values were not fetched and couldn't be displayed.

The reason for this is performance. Executing the edit query once for every attribute value in the list slows down the page load way too much.

So this commit adds another query - the list query - to this individual attribute options.
This allows the operator to provide a query with restrictions using the ci id for editing and provide a separate query which is executed once for the list view without those restrictions.

The new list query is optional.
When empty (like for all existing attributes) the edit query will be used to retain backwards compatibility.